Chitraxi Porwal/requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
requests==2.32.4
bcrypt==4.0.1
cloudant==2.15.0
